requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncio==4.0.0
asyncpg==0.30.0
attrs==25.3.0
boto3==1.40.6
botocore==1.40.6
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
click==8.2.1
coverage==7.10.1
debugpy==1.8.15
distlib==0.4.0
disutils==1.4.32.post2
factory_boy==3.3.3
Faker==37.4.2
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
identify==2.6.12
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==3.0.2
minio==7.2.16
multidict==6.6.3
mypy==1.17.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest_async==0.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.32.0
yarl==1.20.1
